Front End challenge. Card and responsive screen.
This is a simple responsive card build using a container. And inside this one I built a card with the content inside. One first difficulty was put the dice in the middle of the card. Then I used grid, align and justify content. To the responsive mode, I built media queries.